EnKaksi: damn it, that master merge really messed up more than I thought, after clean build something really weird is going on and now I REALLY need firstdiff

+++ include/functions.h
@@ -2889,7 +2889,7 @@ void func_80151BB4(GlobalContext* globalCtx, u32 uParm2);
 // void func_80151DA4(void);
 void func_80152434(GlobalContext* globalCtx, u16 arg2);
 // void func_80152464(void);
-s32 func_80152498(MessageContext* msgCtx);
+u8 func_80152498(MessageContext* msgCtx); // someone else wrote s32 for thiss one
 // void func_8015268C(void);
 // void func_80152C64(void);
 // void func_80152CAC(void);
